% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk.R
\name{VtkVersion}
\alias{VtkVersion}
\title{VTK version used by this package}
\usage{
VtkVersion()
}
\value{
A character string with the VTK version, e.g. \code{"9.3.1"}.
}
\description{
VTK version used by this package
}
